alarm clock: improve efficiency
Problem: We were scheduling waker constantly because we
were not tracking when we next actually needed to run it.

Solution:
- Sort the lists in the sleeping_list priority queue by
wake_me_at, shortest to longest.
- Add a val field to the priority_queue struct
- In sleeping_list, use this val to track the minimum next
time we need to run (i.e. the smallest of the wake_me_at values
in sleeping_list)
- In timer_interrupt, check this value against ticks to see
if we need to up the semaphore