P2: Break up sync APIs into set/signal, misc bugfixing
The child side of "Parent-child IPC" is now done in two
stages: first set the value, then signal. This allows us
to set the value in a few places and signal in exactly one place.

Really only necessary for exit_status, but I did load_status too
for symmetry.

---

Fix a few bugs. Finish implementing syscalls.

It remains to make sure it passes all of the tests.
I think all of the needed code for a working solution is present.