oom-nowait: Add mode where children exit before parent
Not done debugging. It seems to go into a hung state but I
think it's just an "overloaded" system (without a sleep syscall,
children have to spin).

The current implementation has chilren waiting on the creation
of a signal file before they exit. My suspicion is that open ()
is more expensive than write (), so to avoid the hang I think
we can change to denying writes to the signal file.
This avoids the amount of file creation and deletion, which
is a nice bonus.