P2: Add new test rox-fd-leak
This test closes fds from -4096 to 4097 inclusive, excluding
0 and 1 (stdout and stderr).

It then tries to modify its executable.

The purpose is check for "fd leaks" -- if the process is able
to close the fd protecting its own executable from being writable,
this is a leak of resources from the kernel to the user process.

This test is modeled after rox-simple.c.