P2: Fix fd leak discovered by rox-leak-fd
Problem: The fd protecting the file from being modified
(set in load()) was being stored in the thread's fd_table.
As a result, the user could close() and legally close this file.

Solution: Add a separate field to struct thread to represent the
open executable. This way it's protected. No resource leakage.