P3: API and implementation of "process" mmap perspective
These APIs allow you to create an entry in a thread's mmap_table
using an fd.

Intended to be called as part of handling for
syscall_mmap and syscall_munmap

  mapid_t process_mmap_add (int fd);
  void process_mmap_remove (mapid_t);
  struct file * process_mmap_lookup (mapid_t);
  void process_mmap_remove_all (void);

Also includes refactoring of file_table to make it more generic.
The previous version was fd-specific, but now we have both
an fd_table and an mmap_table.

This API needs to be reviewed.