Swap: Address TODOs and general code standard issues
Code not yet tested, of course, but...

- Ensure uniform indentation (2 spaces, as in the rest of Pintos)
- Change API naming convention to match Pintos approach
- Define public interfaces for swap table
- Adjust locking as discussed
- Add more asserts to catch internal errors more promptly.
  Useful during debugging!

The diff looks like I re-did everything, but that's largely because
of changes in indentation.

NB: If your 'tab' key is not set to replace tabs with 2 spaces, please
do so or stop using 'tab'. I can show you the vim setting if you
use vim...