frame table, swap table: Add n_free_entries field.
This field lets us quickly determine whether or not we have
any free frames, slots in the frame, swap table.

If we use a kswapd-style frame eviction policy, this will
facilitate that.

If we need to find a free frame/slot, this makes it easy to
say "Oh, no frames, sorry" without needing to scan the
usage bitmap.