P3: Define a vector based on the file_table
As the mmap_table needs to hold something besides a struct file*,
having a more general vector class becomes necessary.

This class is not yet tested. By converting the file_table code
to use a vector and running the existing file tests from P2,
the class should be exercised enough to be confident that it works.