page: first draft of SPT implementation
Attempts to support shared and non-shared mappings.
Un-tested. Not carefully reviewed.

I'm confident that this will not work on the first try, but it should
be the brunt of what is needed.

Interesting changes that you should look at:

1. 'struct segment' and 'struct shared_mappings' now both use a
'struct segment_mapping_info' which contains:
  - the hash mapping 'relative page number' to page
  - struct file *mmap_file
  - int flags
For a shared mapping, all processes use the same file*.

  Observe that this is not a security breach because each process
needs to be able to open the file in order to give us a file* to
establish the mapping in the first place.

  This allows us to have ONE file instance for a shared mapping, no
matter how many processes are using it. This should reduce memory
consumption and internal load on the file system.

Some tweaks will be required in frame.c and swap.c to make this play nicely.
Presumably whatever code is used to load pages from an mmap'd file will
need to be changed.

2. The stack grows down, so check out segment_calc_page_num
to see how relative page numbers can be calculated when the value of
segment.start is changing.

3. See suggested algorithm change in frame_table_get_eviction_victim