list: 3 new APIs for ordered lists
bool list_insert_ordered_unique: insert elem if not already present
  returns true on insertion

struct list_elem * list_remove_ordered: remove elem from the
  list, returns NULL if no such elem

bool list_contains: is a "copy" of elem present in the list?

Useful for managing the 'struct list owners' of a page.
Were this list expected to be long, using a hash instead of a
list would probably be a good idea...