page: pagedir APIs for use by frame
These APIs take a locked page and iterate over the list of owners
doing the appropriate action.

These two are obvious:
  void page_clear_owners_pagedir (struct page *);
  void page_update_owners_pagedir (struct page *, void *);

This one returns true if the page is dirty. It also sets the
page as clean, since the caller is about to write it to disk if
it is dirty. No sense in traversing the list twice.
  bool page_unset_dirty (struct page *);