stack growth: tweak the behavior when access is just below stack
The logic was a bit confused. The behavior we want is:

In page fault, if it's just below the stack, grow the stack.

Also, update the min observed stack pointer to the faulting address
(so that a subsequent access to the same address after it has been
paged out will not cause spurious stack growth)

In spt_find_segment, define a 'min addr' as seg->start for most
segments and min_observed_stack_pointer for the stack segment.

Unclear if the extra logic in spt_find_segment is now needed,
added a TODO to ponder it.

21 tests remaining.