frame: Accounting of n_free_frames
Problem: In frame_table_evict_page_from_frame, we might have
a FRAME_EMPTY frame. This happens if another process has released
the frame in between find_free_frame and get_eviction_victim.
When it releases the frame, that process incr_n_free_frames, so
we're using a frame that was marked free. In this
case we have to decr_n_free_frames correspondingly.

Also remove some comments and unused code.