cache: implement readahead functionality
This is implemented as a producer-consumer model with multiple producers
and one consumer.

API cache_readahead: Request that the data at the specified address of the
  specified type be asynchronously cached.

Implementation details: This request is expressed as a dynamically allocated
struct that is appended to the readahead_queue (a struct list).
After pushing, a sema is up'd to indicate that there is a new item
available.

The readahead thread sema down's; if this succeeds it locks the queue
and dequeues the first item from the list, requesting that the
item be cached.