cache: fix race conditions
AFAIK The buffer cache is now working.

Summary of outstanding TODOs:
  - Consider making functions for a few misc tasks
  - Revisit replacement algorithm: favor replacing empty blocks and data blocks
  - Revisit what happens when no free blocks are available to evict.
    Perhaps enable the kernel panic rather than cond_wait'ing, at least
    while modifying inode.c.

This is a follow-up to fb1d7d93e.
I deemed the existing state of affairs inadequate because some
of the VM tests were failing; this meant that Dr. Back's tests
were able to hit whatever the issue in the code was.

I believe the largest change was the introduction of rudimentary
synchronization between cache_discard and readahead. The racy
nature of the interaction led me to have cache_discard remove all
pending matching requests and then wait for readahead to finish
its current request, "whatever it was".

This could be improved if cache_discard could see what readahead
was working on (to see if it was the same block or not).
However, this situation should be rare, and the solution is not
hugely expensive since a given readahead operation is not too
expensive. Consequently I am not too concerned about it.

At the moment one test unexpectedly fails: exec-sharing.
This test launches many processes running the same executable.
This means there are many open file*'s to the same underlying inode,
with associated IO through the file.h APIs.
The rest of the failing tests are in the filesys/extended suite,
and are not expected to work.

My suspicion is that exec-sharing is running into an issue with
the implementation of file.c or inode.c, and expect that this test
will start working once we implement synchronization in those areas.

In other words, I think the failure of exec-sharing is actually tied
to the change in filesys_lock and filesys_unlock (specifically
that those functions are now no-ops).

I might be wrong, but since cache_self_test_parallel succeeds
I'm guessing the bug is elsewhere.