cache: fix race on starving process wakeup
Problem: When cond_signal is run, the woken process is scheduled
but not run. Another process could come in and claim the cache block.
This could happen perpetually, causing the woken process to starve.

Solution: Add a boolean is_starving_process to the cache block
to cause other processes to wait on polite_processes for it to
wake up and take its turn.

There is a remaining assert being triggered by exec-sharing in
cache_put_block. As there is no synchronization in inode.c, I
believe that this is the result of competing inode close and open
commands. Not sure, but I'm a loss for an alternative explanation.